requirements.txt
streamlit
google-cloud-firestore
watchdog